Adding bootloader project. Shrinking image size.
Added an initial bootloader project that will eventually perform CRC
verification of app images.

I found that GCC wasn't using nano-newlib as the stdlib implementation,
so I tweaked the config to force it to use that version. This dropped
about 1KB off the bootloader image size and about 3KB off the App image.